arm: snapdragon: Move clock-snapdragon.h to include/
In preparation for handling the clock and reset parts of the clock
controller blocks (yes, I know, they're named 'clock controllers' but
do much more, it wasn't me who chose the name!) without having to
resort to creating two extra fake devices spanning the exact same
register space, make the necessary header public.

While at it, rename "snapdragon" to "qcom" as it's more appropriate
and remove the out-of-place comment in the header.

Note that its relocation in the files that include it is intentional
and will be necessary for subsequent patches.

Signed-off-by: Konrad Dybcio <konrad.dybcio@linaro.org>