dts: sdm845: support the OnePlus 6 and SHIFT6mq
Add support for the OnePlus 6 and 6T devices, as well as the SHIFT6mq.
These are production devices and are based on the dragonboard845c
target. Adjust the dts Makefile to build sdm845 boards when
CONFIG_SDM845 is enabled.

These devices run u-boot as a chain-loaded bootloader and can boot Linux
via GRUB from a fat32 ESP partition.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>