arm: snapdragon: sdm845: simplify init
The KASLR seed can be pulled from the /chosen node in the FDT passed
through by the previous bootloader. This code doesn't work on the boards
I've tested on, remove it for now.

Also remove the button detection, this ought to be handled in a script
using the button API.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>