clk/qcom: add gdsc_enable helper
Add a helper function to enable a GDSC (Global Distributed Switch
Controller). These are responsible for controlling power and clocks to
an entire peripheral. They are represented usually as power domains, but
as they can be enabled with a single register write and live under the
global clock controller, we can just turn them on when the relevent
clock is requested.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>