bootmenu: add an option to loop the bootmenu
On a device with no keyboard, it may be desirable to always loop the
bootmenu and never drop to a shell. This allows the bootmenu to provide
useful options like switching to fastboot or mass storage modes without
then dropping to a console and leaving the device potentially unusable
until it's rebooted. This can be used in tandem AUTOBOOT_USE_MENUKEY to
trigger a menu when a specific key is pressed during boot.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>