mach-snapdragon: add generic Qualcomm target
The board-specific targets don't scale particularly well, and largely
exist to handle board-specific configuration which could not easily be
done in a generic way. However, changes in this and future series will
allow us to replace most of the board specific targets with a single
one.

Introduce a generic Qualcomm target, to eventually encompass all
Qualcomm board support in u-boot. This target expects the devicetree to
be passed in by the previous stage bootloader, this populates the memory
node and allows us to dynamically generate the memory map.

This target currently supports the dragonboard845c and starqltechn, both
of which will be removed in favour of this common target.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>