mach-snapdragon: drop dragonboard845c target
The db845c is fully supported by the generic Qualcomm target, remove the
board specific target in favour of that. This requires slightly
adjusting how u-boot is compiled, now instead of the DTB being built
into u-boot with a dummy DTB appended on, the u-boot DTB can be appended
as normal so that ABL will populate the memory node. u-boot will then
retrieve the DTB address and use it as the primary DT.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>