mach-snapdragon: drop starqltechn target
Like the db845c, starqltechn is also fully supported by passing the DTB
via ABL. The KASLR address can also be read directly from the DT as
and passed through to Linux. The button handling is dropped as the same
behaviour can be achieved through the autoboot configuration. Drop the
device-specific target as well the sdm845 sysmap and custom init code.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>