mach-snapdragon: drop qcs404-evb target
The QCS404 EVB can be fully supported by the generic Qualcomm target.
The only board specific code was to enable the USB VBUS GPIO, this is
now enabled via the dwc3 generic glue using a GPIO regulator defined in
DTS.

Enable the QCS404 clock and pinctrl drivers in qcom_defconfig.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>