doc: board/qualcomm: document generic targets
Replace the board specific docs with a generic board.rst file which
documents the build/boot process for the sdm845 and qcs404 boards now
that the only differences are the DTB in use.

At the same time, create a debugging page to document some useful
snippets and tips for working with Qualcomm platforms.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>