clk/qcom: add gdsc_enable helper
Global Distributed Switch Controllers are per-domain switches which are
used to toggle power and clocks to an entire subsystem. They live under
the GCC block and might need to be enabled before certain clocks, so
handle them as part of the clock driver.

Linux models these as power domains, however this additional complexity
doesn't offer much benefit to us in U-Boot. For now they can be turned
on as-needed when a relevant clock is enabled.

In the future, we can add a power-domain driver to model these properly.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>