Qualcomm preliminary USB support
This series introduces support for Qualcomms DWC3 controller glue, as
well as two of the high speed PHYs found on various Qualcomm platforms
and suport for programming the SMMU to enable DMA for the USB
controller.

Qualcomm platforms are heavily locked down, in many cases offering no
way for users to gain EL2 or EL3 code execution and including many
additional security features. Among these are modifications made to
the SMMU behaviour (done by having the hypervisor intercept register
accesses). On many platforms it is not possible to disable the SMMU,
and not possible to configure streams in BYPASS.

To enable support for peripherals like USB which depend on DMA, we must
allocate a context bank on the SMMU and configure an IDENTITY mapping.

This patch series fixes support for building IOMMU without CONFIG_PCI.

Then, it introduces first introduces support for a .connect() op in
IOMMU drivers, this is called before the peripheral drivers .probe()
function and allows for IOMMU devices to perform one-time setup.

Then, introduce a new driver which supports the SMMU as found on
Qualcomm platforms, when a DMA capable peripherals is probed, this
driver will first see if the previous stage bootloader configured a
context bank for the given stream ID already, and if not then it
allocates and configures one so that DMA accesses will work.

Additionally, it adds the necessary clock driver code to bring up both
USB controllers on SDM845.

The USB CDC/ACM serial gadget has a bug where it doesn't initialise the
USB controller automatically. This is fixed by introducing a call to
usb_gadget_initialize().

Lastly, the mass storage gadget driver gains support for multiple sector
sizes as the UFS storage on Qualcomm devices uses a non-standard 4096
sector size. The changes are based on a similar patch to the Linux
version of the driver.

This series replaces my previous one sending only the SMMU driver

https://lore.kernel.org/u-boot/20231211-b4-dwc3-qcom-v1-0-46275113b4f2@linaro.org/

This series can be applied standalone, but has some dependencies for
testing. A feature branch based on qcom-next with the necessary
dependencies for testing the Dragonboard845c can be found here:

https://git.codelinaro.org/linaro/qcomlt/u-boot/-/tree/b4/dwc3-qcom

To: Caleb Connolly <caleb.connolly@linaro.org>
To: Neil Armstrong <neil.armstrong@linaro.org>
To: Sumit Garg <sumit.garg@linaro.org>
To: Lukasz Majewski <lukma@denx.de>
To: Sean Anderson <seanga2@gmail.com>
To: Marek Vasut <marex@denx.de>
Cc:  <u-boot@lists.denx.de>

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 1,
    "change-id": "20231212-b4-dwc3-qcom-a95e436160fb",
    "prefixes": []
  }
}