board/qualcomm: add config fragments for uboot-in-XBL
Running U-Boot as the primary bootloader on Qualcomm platforms requires
setting board specific options to bake in the DTS and build an ELF file
with a specific TEXT_BASE. Create fragments for the rb2 and rb3 (db845c)
which configure this (and include some disabled debug options to make
development easier).

With this, the resulting u-boot.elf can be packed into an XBL partition
image using $METHOD and booted.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>