mach-snapdragon: prioritse internal DTB
If U-Boot has a DTB built in (appended to the image directly) then this
was likely intentional, we should prioritise it over one provided by ABL
(if there was one).

Make this behaviour explicit, and panic if no valid DTB could be found
anywhere. Returning an error is not useful in this case.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>