mach-snapdragon: use U-Boot FDT as fallback
We pre-allocate a memory region for the FDT, and as a result U-Boot will
always expose the EFI DT protocol - even if the region is empty. Rather
than solely relying on the ESP containing a DTB file at the right path,
let's play it super safe and copy U-Boots FDT in as a fallback.

This way, even if no FDT is loaded the kernel will at least be able to
output to the serial port.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>