Qualcomm platform USB support
This series enables USB on Qualcomm SDM845 platforms and lays the
foundation for future SoCs as well.

It introduces two new high-speed PHY drivers, one for SDM845 and one for
an upcoming platform. The SDM845 clock driver gains support for
configuring the USB clocks, and the GPIO driver is updated to use .set_flags
which fixes a strange bug where GPIOs would also be configured as input.

Support for super-speed USB modes is not currently available, however
configuring the device to be high-speed only requires modifications to
DT.

To improve compatibility with upstream DT, we switch Qualcomm platforms
over to OF_LIVE and apply fixups to the tree to remove references to the
super-speed phy, as well as configure the Qualcomm glue for high-speed
only mode.

The db845c requires a vbus-supply to be hooked up for its usb type-A
port. A U-Boot dtsi file is added to configure this.

With these patches, it is now also possible to run U-Boot on some SDM845
phones like the OnePlus 6 using a dtb from Linux, and access the U-Boot
shell via the CDC ACM USB serial gadget.

This series depends on the ("Qualcomm generic board support") series [1]
switching to upstream DT, as well as the ("Qualcomm DWC3 USB support")
series [2]. A feature branch based on qcom-next with the necessary
dependencies for testing the Dragonboard845c can be found at [3].

I think most of these can go through the Qualcomm tree, but I'm not
exactly sure. I'd appreciate any suggestions on if/how to split this
series up further.

[1]: https://lore.kernel.org/u-boot/20240130-b4-qcom-common-target-v3-0-e523cbf9e556@linaro.org/
[2]: https://lore.kernel.org/u-boot/20240131-b4-qcom-usb-v1-0-6438b2a2285e@linaro.org/
[3]: https://git.codelinaro.org/linaro/qcomlt/u-boot/-/tree/b4/dwc3-qcom

To: Tom Rini <trini@konsulko.com>
To: Caleb Connolly <caleb.connolly@linaro.org>
To: Neil Armstrong <neil.armstrong@linaro.org>
To: Sumit Garg <sumit.garg@linaro.org>
To: Lukasz Majewski <lukma@denx.de>
To: Sean Anderson <seanga2@gmail.com>
Cc:  <u-boot@lists.denx.de>

---
Changes in v2:
- Fix incorrect order of NULL/0 parameters to of_write_prob() in
  fixup_qcom_dwc3()
- Add missing break to switch case in sdm845_clk_enable()
- Remove rogue return statement in msm_gpio_set_flags()
- Link to v1: https://lore.kernel.org/r/20240131-b4-qcom-livetree-v1-0-4071c0787db0@linaro.org

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 2,
    "change-id": "20240131-b4-qcom-livetree-710b265c695f",
    "prefixes": [],
    "history": {
      "v1": [
        "20240131-b4-qcom-livetree-v1-0-4071c0787db0@linaro.org"
      ]
    }
  }
}