WIP: arm: dts: add OnePlus 8T
The OnePlus 8T (oneplus-kebab) is an SM8250 based flagship phone from
2020. It has been imbued with signing key fingerprints belonging to
Qualcomm and OnePlus, as such the Android bootloader stack cannot be
replaced. As a result, U-Boot can only be run via chainloading, where
it's wrapped up in an Android boot image and jumped to from the Android
bootloader.

Initial support includes the display (via framebuffer), buttons, UFS
storage and USB type-c gadget mode.

This DT is not yet in Linux.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>