ufs: qcom: initialise in G3/FAST_MODE
We don't yet have a nice way to get the max clk frequency (should parse
from freq-table-hz in DT), but all platforms are either 150 or 300MHz.
We can just program in the timings for 300MHz and it doesn't seem to
break SDM845 - but more testing needed.

This fixes UFS being extremely slow and buggy on SM8250 due to being
stuck at the slowest speed with misconfigured timings.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>