fs: fat: allow FS and device to have mismatching sector sizes
The Qualcomm Robotics RB1 and RB2 boards feature MMC storage provisioned
with a 4k sector size GPT, with factory partitions formatted with
filesystems using 512 byte sector sizes.

Whilst certainly unusual, and really not ideal, this appears to work
fine in practise. So demote this failure condition to a debug log.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>