qcom_defconfig: set a safe CONFIG_SYS_LOAD_ADDR
This was previously set to 0, however certain load/boot commands end up
trying to read from this address when run with no arguments, causing an
abort on all Qualcomm platforms.

Instead of this highly undesirable behaviour, let's pick an address
which is valid unused RAM on all the Qualcomm boards supported so far.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>