dm: core: add of_find_node_by_prop()
We have of_find_node_by_prop_value() which allows searching for a node
by matching both the property name and value. Introduce a new helper to
find a node only by property, and return the property as well.

This is useful in cases where we want to modify the live tree, for
example to remap volume keys to up/down by searching for every node with
a "linux,code" property and then operating on its value.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>