efi: add a helper to generate dynamic UUIDs
Introduce a new helper efi_capsule_update_info_gen_ids() which takes
several strings to identify the currently running board as well as a
platform specific salt UUID and uses this data to populate the capsule
update fw images image_type_id field. This allows for determinstic UUIDs
to be used that can scale to a large number of different boards and
board variants without the need to maintain a big list.

Generating capsule updates can be done using the same namespace, soc,
model, compatible, and fw_image name strings.

This is behind an additional config option as it depends on V5 UUIDs and
the SHA1 implementation.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>