cli: enable bootretry for hush
We can retry boot just by copying the command to run into the
console_buffer. Do this with "run bootcmd", enabling bootretry for hush.
A nice usecase for this is to make hush repeatedly launch a bootmenu,
allowing for a usable UX to be provided for devices like phones without
serial or a keyboard.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>