usb: dwc3: respect role-switch-default-mode
* Factor out the common pattern of checking the dr_mode property on
  the usb node and it's parent
* Respect the usb-role-switch property, rather than requiring dr_mode be
  set
* Override OTG mode with the value in role-switch-default-mode

The devicetree bindings don't dictate that dr_mode is a required
property, but the dwc3 driver would refuse to probe if it was unset
(this breaks the Qualcomm RB2 board which doesn't set it).

Here, we update the driver to respect the other properties that can be
used to describe the controller, and more gracefully handle some of the
edge cases.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>