mach-snapdragon: support parsing memory map from SMEM
It is possible to derive the memory map for a Qualcomm platform from the
SMEM shared memory region. The memory map is populated by the preloader.

Introduce support for parsing this data and using it to populate
U-Boot's memory map. Since we aren't yet sure if this will work for
every platform, it is not yet used in all cases, if U-Boot is booted
with an internal FDT which has the memory map defined, this will
be used instead.

If the FDT comes from ABL, or we're using an internal FDT with no memory
map defined, then U-Boot will try to use SMEM. This should remove the
need to define the memory map statically in a U-Boot overlay DT for most
boards.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>