usb: dwc3: bind dwc3 host and gadget
The dwc3 generic driver support host and gadget modes, with separate
drivers for each. Both initialise the dwc3 core in the expected mode and
expose the relevant interface (XHCI or UDC).

Currently, only one of these drivers is bound depending on the DT
properties for the device, with OTG controllers being treated as
peripheral mode.

However, there's no reason we can't actually bind both drivers, just so
long as we ensure they aren't both probed at the same time.

Implement support for this, using the child_pre_probe callback on the
wrapper to remove the host device when probing the gadget and vice
versa, ensuring they remain mutually exclusive.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>
---
Tested on a few Qualcomm platforms, this makes it possible to seamlessly
switch between gadget mode and host mode on the same controller as
needed.