qcom: smem: modernize SMEM in U-Boot
Smem is currently implemented a uclass in U-Boot. This carries with it
some implicit limitations about how and when we can actually talk with
it, since we must allocate an smem device "handle".

Since smem is /just/ a special memory region, and specific to Qualcomm
platforms, trying to coerce it into some standard interface only makes
using it more difficult.

Let's drop the old smem code and replace it with a straight port from
Linux, providing a familiar interface and enabling us to access it much
easier in the boot process.

With this new smem framework in place, we can now use SMEM to parse
the memory map when it's not provided via some other mechanism.
Additionally, we can use it to read the serial number of the board.

Implement this functionality and split the snapdragon memory parsing off
to its own file to try and organise things a bit better.

To: Rayagonda Kokatanur <rayagonda.kokatanur@broadcom.com>
To: Tom Rini <trini@konsulko.com>
To: Simon Glass <sjg@chromium.org>
To: Caleb Connolly <caleb.connolly@linaro.org>
To: Neil Armstrong <neil.armstrong@linaro.org>
To: Sumit Garg <sumit.garg@linaro.org>
To: Mario Six <mario.six@gdsys.cc>
Cc: u-boot@lists.denx.de
Cc: u-boot-qcom@groups.io


--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 1,
    "change-id": "20241121-b4-modernise-smem-b456b8e21d6b",
    "prefixes": []
  }
}