[HACK]bootm: ARM64 EL1 direct jump to kernel
It was EL3->EL2 or EL2->EL1 when boot. But in many case such chain
loading from bootloader run in EL1. It is assumed practically no
way to raise EL1 to EL2 ;-)

Thus the only way to start kernel is direct jump. This apporach is
similar to KEXEC purgatory, and the low level code is borrow fro it.

Signed-off-by: George Chan <gchan9527@gmail.com>