usb: dwc3: gadget: conditionally remove requests Date: Wed, 20 Jul 2022 23:35:23 +0200	[thread overview] Message-ID: <20220720213523.1055897-1-m.grzeschik@pengutronix.de> (raw)
The functions stop_active_transfers and ep_disable are both calling
remove_requests. This functions in both cases will giveback the requests
with status ESHUTDOWN, which also represents an physical disconnection.
For ep_disable this is not true. This patch adds the status parameter to
remove_requests and sets the status to ECONNRESET on ep_disable.

Signed-off-by: Michael Grzeschik <m.grzeschik@pengutronix.de>