dts: sm7125: ufs: Disable ufs completely
To play safe, disable ufs access to u-boot. Such as fastboot flash cant
access to, or chainload to, any of ufs partitions.

To chainload original ufs->boot partition, there is a workaround to
flash it in a mmc boot partition, then chainload to it. Fastboot boot
cmd does not affect by this change.

Signed-off-by: George Chan <gchan9527@gmail.com>